+++ test/fixtures/withServerContext-page-with-getInitialProps/pages/index.js
@@ -1,7 +1,5 @@
-'use strict';
-
-const useServerContext = require('../../../../public/useServerContext.js');
-const withServerContext = require('../../../../public/withServerContext.js');
+import useServerContext from '../../../../public/useServerContext.mjs';
+import withServerContext from '../../../../public/withServerContext.mjs';
 
 function IndexPage({ pageCustomProp }) {
   const serverContext = useServerContext();
@@ -22,4 +20,4 @@ IndexPage.getInitialProps = async () => ({
   pageCustomProp: 'pageCustomProp' + '_value',
 });
 
-module.exports = withServerContext(IndexPage);
+export default withServerContext(IndexPage);